More changes to the player classes and the team class

+++ src/edu/gsu/csc1302/emperorsofspades/team/Team.java
@@ -96,7 +96,7 @@ public class Team {
      *
      * @return
      */
-    public int placeTeamBid() {
+    public double placeTeamBid() {
         return this.teammates.get(1).placeBid();
     }
